Update request methods to current API requirements
The library previously sent translation/detection POST requests with a URL-encoded form submission body — though without the corresponding `Content-Type: application/x-www-form-urlencoded` header. It looks like this was a strategy to maximise the amount of text that could be sent in a single request — from an old version of the Google Translate API docs, clients could previously either pass up to 2k characters of text in a `q` query parameter with a GET request, or up to 5k characters if sent in the body of a POST request that set the header `X-HTTP-Method-Override: GET`.

The Google Translate API v2 seems to have previously tolerated the lack of a content type header — inferring it from the body — but at some point in the fall of 2025 the API started responding to these requests in the following way:
```shell
< HTTP/2 400
< vary: X-Origin
< vary: Referer
< vary: Origin,Accept-Encoding
< content-type: application/json; charset=UTF-8
< date: Wed, 05 Nov 2025 11:37:57 GMT
< server: ESF
< x-xss-protection: 0
< x-frame-options: SAMEORIGIN
< x-content-type-options: nosniff
< alt-svc: h3=":443"; ma=2592000,h3-29=":443"; ma=2592000
< accept-ranges: none
<
{
  "error": {
    "code": 400,
    "message": "Invalid JSON payload received. Unexpected token.\nq=hello+world\n^",
    "errors": [
      {
        "message": "Invalid JSON payload received. Unexpected token.\nq=hello+world\n^",
        "domain": "global",
        "reason": "parseError"
      }
    ],
    "status": "INVALID_ARGUMENT"
  }
}
```

From an experiment, requests with a URL-encoded form submission body do still succeed if a `Content-Type: application/x-www-form-urlencoded` header is provided.

However, given that today's v2 docs only mention support for JSON requests — presumably anything else lacks official support — this commit switches to the JSON-based format that is recommended in 2025.

Also removes the `X-HTTP-Method-Override: POST` header, which appears to no longer be necessary if we're departing from the original workaround (and also isn't mentioned in today's docs). Request limits (at least when using a JSON body) are now much higher than when this library was first written: 100k characters per request (see https://docs.cloud.google.com/translate/quotas)

```